requirements.txt
openai>=1.3.9 # OpenAI API client
python-dotenv  # for keeping your API key safe in .env
